backend/requirements.txt
langchain-pinecone
langchain-community
langchain-text-splitters
langchain-tavily
langchain-openai
langgraph
langchain
pypdf
pinecone

fastapi
uvicorn
ipykernel

python-dotenv
openai

faster-whisper
edge-tts
python-multipart
